Add telegraf plan using static binary
This plan builds on the work started by @kplimack in #639.

Telegraf project provides precompiled static binaries.  They're
referenced and linked on the Github project "releases" page rather
than the Influx project portal's "downloads" page for telegraf. The
"downloads" page has only the dynamically linked binary, which has a
libc interpreter outside aa path that Habitat knows about (the
standard /lib location for Linux 64 bit machines).

This plan will make use of the statically compiled binary so we
don't have to:

1. Build from source, or
2. Use patchelf to change the interpreter of the dynamically built
   binary

As telegraf has highly customizable config, it's not maintainable to
template the configuration, so it is recommended users create a
package that depends on this one. This is described in greater detail
in the README.md.

Signed-off-by: Joshua Timberman <joshua@chef.io>